Fix parallel execution bug that only ran connect/preconfigure stuff on the last testbed: lambda : testbed.do_something   only does it for the last value of the variable "testbed"
Testbed is a free variable, so it takes the value of the variable then the lambda is called,
which is usually the last value it took in the loop.

The correct way would be:
lambda testbed=testbed : testbed.do_something

Since that "freezes" the value of the "testbed" variable.

An easier and nicer way is to use bound methods:

testbed.do_something

(without lambda)